test: malformed `DomainRuleNode`
Currently, when creating a DomainRuleNode with a grammar,
it's possible to have a mismatch between the number of children
in each of the rules in the domain and the number of children
supplied to the `DomainRuleNode`. This can and usually does cause
cryptic errors later on about not being about to match nodes with
different arities. This test checks that case explicitly and
ensures that the constructor errors.