fix: check rule arity in `DomainRuleNode`
When constructing a `DomainRuleNode` with a grammar,
we can check whether all of the rules in the domain
have the same arity and whether the number of children
passed to the `DomainRuleNode` constructor match. We
should enforce this here because when these don't match,
we sometimes end up with some errors that are difficult
to understand during constraint propagation.